components/arena/docker/requirements.txt
requests==2.18.4
six==1.11.0
pyyaml==3.12
